requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
python-jose
python-multipart
passlib[bcrypt]
bcrypt
email-validator
pyodbc
python-dotenv
passlib[argon2]